+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 0.7.0
-- Added MemoryStorageConnector (useful for testing)
 
+- Added MemoryStorageConnector (useful for testing)
+- Added `load` method to `CloudWandererResource`
+- Added `load` support to `MemoryStorageConnector`
+- Added `load` support to `DynamoDbConnector`
 
 # 0.6.0